Add the --version command line flag
This adds a new method to the Panda class called `version` which, if in a
git working directory determines the version from the most recent tag via
`git describe` otherwise it tries to extract the version information from
`META.info`.  At present the solution is a bit hacky, however as soon as
information can be extracted from the DISTRIBUTION variable, the
implementation can be improved.  This commit thus closes issue #204 and by
inference also issue #198.