Warn if --no-tests or --no-deps is used
It is common for people to mistakenly use --no-tests or --no-deps as options
to panda: the correct options are --notests and --nodeps, respectively.
This change warns the user about the incorrect options, tells them what the
correct options are, prints the usage information and exits.  It also warns
and exits if it finds options it didn't expect.